Updated create action in foods_controller to make a call to the Favorite model upon instantiation of a new instance of the Food class, to also instantiate a new instance of the Favorite class, based on the newly created food's id and provided params. Updated favorite.rb file in models folder to create a class method, create_favorite, which is called from the creat action in the foods_controller. Updated favorite_params in favorites_controller. Updated meal_params in the meals_controller. Updated user_params in the users_controller. Changed data type of nutritional information data in foods table from type decimal to type float. Updated seeds.rb file and dropped existing database and reseeded with current data.

+++ db/migrate/20210823043355_create_foods.rb
@@ -2,20 +2,20 @@ class CreateFoods < ActiveRecord::Migration[6.1]
   def change
     create_table :foods do |t|
       t.string :food_name
-      t.decimal :calories
-      t.decimal :cholesterol
-      t.decimal :dietary_fiber
-      t.decimal :potassium
-      t.decimal :protein
-      t.decimal :saturated_fat
-      t.decimal :sodium
-      t.decimal :sugars
-      t.decimal :total_carbohydrate
-      t.decimal :total_fat
+      t.float :calories
+      t.float :cholesterol
+      t.float :dietary_fiber
+      t.float :potassium
+      t.float :protein
+      t.float :saturated_fat
+      t.float :sodium
+      t.float :sugars
+      t.float :total_carbohydrate
+      t.float :total_fat
       t.string :thumbnail
-      t.decimal :serving_qty
-      t.decimal :serving_unit
-      t.decimal :serving_weight_grams
+      t.float :serving_qty
+      t.float :serving_unit
+      t.float :serving_weight_grams
 
       t.timestamps
     end
